+++ resources/views/components/settings/layout.blade.php
@@ -18,8 +18,12 @@
     <flux:separator class="md:hidden" />
 
     <div class="flex-1 self-stretch max-md:pt-6">
-        <flux:heading>{{ $heading ?? '' }}</flux:heading>
-        <flux:subheading>{{ $subheading ?? '' }}</flux:subheading>
+        <x-daisyui.heading class="mb-2">{{ $heading ?? ''}}</x-daisyui.heading>
+        
+        <x-daisyui.heading type="sub">{{ $subheading ?? ''}}</x-daisyui.heading>
+
+        {{-- <div class="text-lg">text-lg {{ $heading ?? '' }}</div>
+        <div class="text-xl">text-xl {{ $heading ?? '' }}</div> --}}
 
         <div class="mt-5 w-full max-w-lg">
             {{ $slot }}