+++ resources/views/livewire/settings/delete-user-form.blade.php
@@ -24,8 +24,8 @@ new class extends Component {
 
 <section class="mt-10 space-y-6">
     <div class="relative mb-5">
-        <flux:heading>{{ __('Delete account') }}</flux:heading>
-        <flux:subheading>{{ __('Delete your account and all of its resources') }}</flux:subheading>
+        <x-daisyui.heading class="mb-2">{{ __('Delete account') }}</x-daisyui.heading>
+        <x-daisyui.heading type="sub">{{ __('Delete your account and all of its resources') }}</x-daisyui.subheading>
     </div>
 
     <flux:modal.trigger name="confirm-user-deletion">
@@ -38,11 +38,11 @@ new class extends Component {
     <flux:modal name="confirm-user-deletion" :show="$errors->isNotEmpty()" focusable class="max-w-lg">
         <form wire:submit="deleteUser" class="space-y-6">
             <div>
-                <flux:heading size="lg">{{ __('Are you sure you want to delete your account?') }}</flux:heading>
+                <x-daisyui.heading size="lg" class="mb-2">{{ __('Are you sure you want to delete your account?') }}</x-daisyui.heading>
 
-                <flux:subheading>
+                <x-daisyui.heading type="sub">
                     {{ __('Once your account is deleted, all of its resources and data will be permanently deleted. Please enter your password to confirm you would like to permanently delete your account.') }}
-                </flux:subheading>
+                </x-daisyui.heading>
             </div>
 
             <flux:input wire:model="password" id="password" label="{{ __('Password') }}" type="password" name="password" />