+++ resources/views/partials/settings-heading.blade.php
@@ -1,5 +1,5 @@
 <div class="relative mb-6 w-full">
-    <flux:heading size="xl" level="1">Settings</flux:heading>
-    <flux:subheading size="lg" class="mb-6">{{ __('Manage your profile and account settings') }}</flux:subheading>
+    <x-daisyui.heading size="xl" level="1">Settings</x-daisyui.heading>
+    <x-daisyui.heading type="sub" size="lg" class="mb-6">{{ __('Manage your profile and account settings') }}</x-daisyui.heading>
     <flux:separator variant="subtle" />
 </div>